hedge forms default values
